Added calibration program
During calibration the user will indicate which area of the image is the
scrumboard and will select a part of the scrumboard that will never be
obscured. This selected part will be used to remove color casts by
always correcting it to a shade of grey. Calibration data will be
written to calibrationdata.json