findnotes: try to determine note center based on mode filter
The problem I had with determining the center purely based on the
coloredness of the pixels is that the shadow of a note would also be
colored, thereby possibly incorrectly finding the center offset towards
the shadow.

The mode-based solution does not seem to suffer from this problem with
shadow, but there is another problem. The mode in an off-center part of
the note may have a higher occurrence than the mode in the center. My
theory is that this happens when the first is darker, leading to more
matches with the pixels around the black text.