Added the beginning of a web interface based on twisted nevow
The web interface now shares access to the webcam with the tracker, both
of which are running in a twisted event loop, making it easy to avoid
simultaneous access to the webcam.
The tracker saves the image into the scrumboard state and passes that to
the boardreader that it launches. This means that the boardreader itself
no longer accesses the webcam.